+++ src/orm/bib_volume.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import ForeignKey, Integer
+from sqlalchemy import ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 import src.orm.base
@@ -12,7 +12,7 @@ class BibVolume(src.orm.base.Base):
         'info': dict(is_view=True),
         'schema': 'sierra_view'
     }
-    id: Mapped[int] = mapped_column("id", Integer, primary_key=True)
+    id: Mapped[int] = mapped_column("id", BigInteger, primary_key=True)
     bib_record_id: Mapped[int] = mapped_column("bib_record_id", ForeignKey("sierra_view.bib_record.record_id"))
     volumes_record_id: Mapped[int] = mapped_column("volume_record_id", ForeignKey("sierra_view.volume_record.record_id"))
     volumes_display_order: Mapped[int] = mapped_column("volumes_display_order")
