+++ src/orm/volume.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List
 
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 import src.orm.base
@@ -18,8 +18,8 @@ class Volume(src.orm.base.Base):
         'info': dict(is_view=True),
         'schema': 'sierra_view'
     }
-    id: Mapped[int] = mapped_column("id", Integer, primary_key=True)
-    record_id: Mapped[int] = mapped_column("record_id", Integer, primary_key=True)
+    id: Mapped[int] = mapped_column("id", BigInteger, primary_key=True)
+    record_id: Mapped[int] = mapped_column("record_id", BigInteger, primary_key=True)
     sort_order: Mapped[int] = mapped_column("sort_order")
     is_suppressed: Mapped[bool] = mapped_column("is_suppressed")
     bibs: Mapped[List["src.orm.bib.Bib"]] = relationship(
